// typescript/sdk/src/cw-types/IsmMultisig.types.ts
/**
 * This file was automatically generated by @cosmwasm/ts-codegen@0.35.3.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run the @cosmwasm/ts-codegen generate command to regenerate this file.
 */

export interface InstantiateMsg {
  owner: string;
}
export type ExecuteMsg =
  | {
      ownable: OwnableMsg;
    }
  | {
      enroll_validator: {
        set: ValidatorSet;
      };
    }
  | {
      enroll_validators: {
        set: ValidatorSet[];
      };
    }
  | {
      unenroll_validator: {
        domain: number;
        validator: HexBinary;
      };
    }
  | {
      set_threshold: {
        set: ThresholdSet;
      };
    }
  | {
      set_thresholds: {
        set: ThresholdSet[];
      };
    };
export type OwnableMsg =
  | {
      init_ownership_transfer: {
        next_owner: string;
      };
    }
  | {
      revoke_ownership_transfer: {};
    }
  | {
      claim_ownership: {};
    };
export type HexBinary = string;
export interface ValidatorSet {
  domain: number;
  validator: HexBinary;
}
export interface ThresholdSet {
  domain: number;
  threshold: number;
}
export type QueryMsg =
  | {
      ownable: OwnableQueryMsg;
    }
  | {
      ism: IsmQueryMsg;
    }
  | {
      multisig_ism: MultisigIsmQueryMsg;
    };
export type OwnableQueryMsg =
  | {
      get_owner: {};
    }
  | {
      get_pending_owner: {};
    };
export type IsmQueryMsg =
  | {
      module_type: {};
    }
  | {
      verify: {
        message: HexBinary;
        metadata: HexBinary;
      };
    }
  | {
      verify_info: {
        message: HexBinary;
      };
    };
export type MultisigIsmQueryMsg = {
  enrolled_validators: {
    domain: number;
  };
};
export interface EnrolledValidatorsResponse {
  threshold: number;
  validators: HexBinary[];
}
export type Addr = string;
export interface OwnerResponse {
  owner: Addr;
}
export interface PendingOwnerResponse {
  pending_owner?: Addr | null;
}
export type IsmType =
  | 'unused'
  | 'routing'
  | 'aggregation'
  | 'legacy_multisig'
  | 'merkle_root_multisig'
  | 'message_id_multisig'
  | 'null'
  | 'ccip_read';
export interface ModuleTypeResponse {
  type: IsmType;
}
export interface VerifyResponse {
  verified: boolean;
}
export interface VerifyInfoResponse {
  threshold: number;
  validators: HexBinary[];
}